% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_prefs.R
\name{backup_prefs}
\alias{backup_prefs}
\title{Backup of R Studio preferences.}
\usage{
backup_prefs(open_prefs_folder = FALSE)
}
\arguments{
\item{open_prefs_folder}{logical: TRUE or FALSE}
}
\value{
Return keybindings, preferences and snippets in .bak in your R Studio
  config path.
}
\description{
Use \code{usethis:::rstudio_config_path()} to obtain your preference path.
}
\examples{
Simply use backup_prefs() to backup your preferences.
Use backup_prefs(open_prefs_folder = TRUE) to access your backup files

}
