% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_local_prefs.R
\name{copy_files_to_local}
\alias{copy_files_to_local}
\title{Copy backup files to a chosen local folder}
\usage{
copy_files_to_local(preference_path_name = "rstudio_preferences")
}
\arguments{
\item{preference_path_name}{string: Use a relative file path, don't add a "/" at the end}
}
\value{
Return addins.json, r.snippets, rstudio_bindings.json, prefs.json in default folder (rstudio-preferences) which can be modified.
}
\description{
Copy backup files to a chosen local folder
}
\examples{
copy_files_to_local()

}
