% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_to_local_prefs.R
\name{files_to_copy}
\alias{files_to_copy}
\title{Internal function usage to get the vector of json files to copy
Useful when there is a file missing such as r.snippets}
\description{
Internal function usage to get the vector of json files to copy
Useful when there is a file missing such as r.snippets
}
