% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_git_repository}
\alias{has_git_repository}
\title{Does it have a valid git repository?}
\usage{
has_git_repository()
}
\description{
Simply check for git status and assert that it is error free. Internal usage.
}
