% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_local_prefs.R
\name{import_local_prefs}
\alias{import_local_prefs}
\title{Import local preferences files to R Studio}
\usage{
import_local_prefs(preference_path = "/rstudio_preferences")
}
\arguments{
\item{preference_path}{string: Relative folder_path where .json files are
contained.}
}
\value{
Import addins.json, rstudio_bindings.json, r.snippets to the
keybindings folder and rstudio-prefs.json to the RStudio folder
}
\description{
Import local preferences files to R Studio
}
\examples{
import_local_prefs("/rstudio_preferences")
}
