% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_git_repositories.R
\name{list_github_repositories}
\alias{list_github_repositories}
\title{List the online git repositories of the current user}
\usage{
list_github_repositories(username = get_current_git_username)
}
\arguments{
\item{username}{string. The git username to list the repositories. Default is
the current git username.}
}
\value{
A list of the current public repositories for the current git user
}
\description{
List the online git repositories of the current user
}
\examples{
list_github_repositories("cran")
\dontrun{
list_github_repositories()
}

}
