% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudioBackupPrefs-package.R
\docType{package}
\name{RStudioBackupPrefs-package}
\alias{RStudioBackupPrefs}
\alias{RStudioBackupPrefs-package}
\title{RStudioBackupPrefs: Easily backup your RStudio preferences easily and import them on another computer from github}
\description{
The package provides a few short functions that automates backup and import from local or github files. Check the documentation for usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Minh-AnhHuynh/RStudioBackupPrefs}
  \item \url{https://minh-anhhuynh.github.io/RStudioBackupPrefs/}
  \item Report bugs at \url{https://github.com/Minh-AnhHuynh/RStudioBackupPrefs/issues}
}

}
\keyword{internal}
