% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_backup_prefs.R
\name{start_backup_prefs}
\alias{start_backup_prefs}
\title{Start the backup process}
\usage{
start_backup_prefs(
  github_backup = FALSE,
  preference_path = "rstudio_preferences"
)
}
\arguments{
\item{github_backup}{boolean: Upload file to github.}

\item{preference_path}{string: Path to local backup folder.}
}
\value{
Copy rstudio_bindings.json, addins.json rstudio-bindings.json and
r.snippets to .bak files in their native config folder. Then copy files to
working directory in path and executes git push if \code{github_backup = TRUE}
}
\description{
Function will start backing up your files in their folder to .bak and ask you
if you want to back them up to GitHub. Make sure to have a git initiated
repository.
}
\details{
Function wrapper using \code{backup_prefs()}. if github_backup = TRUE,
will then use those functions: \code{copy_files_to_local()} and
\code{upload_prefs_to_github()}
}
\examples{
\dontrun{
start_backup_prefs(github_backup = TRUE)
}

}
