% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_prefs_github.R
\name{upload_prefs_to_github}
\alias{upload_prefs_to_github}
\title{Upload preferences to GitHub}
\usage{
upload_prefs_to_github(
  preference_path = "/rstudio_preferences",
  git_message = "Backup of R Studio preferences on {Sys.Date()}."
)
}
\arguments{
\item{preference_path}{string. Relative file path name}

\item{git_message}{string. Enter a git commit message.}
}
\value{
Upload your .json files to currently active git repository.
}
\description{
Add, commit and push your .json files automatically with the name
}
\examples{
\dontrun{
upload_prefs_to_github("/rstudio_preferences")
upload_prefs_to_github("/rstudio_preferences", git_message = "Backup preferences")
}

}
