% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_prefs.R
\name{backup_prefs}
\alias{backup_prefs}
\title{Backup of R Studio preferences.}
\usage{
backup_prefs(
  open_backup_path = FALSE,
  rstudio_pref_path = rstudio_config_path()
)
}
\arguments{
\item{open_backup_path}{logical: TRUE or FALSE. Open the preference folder
at the end of backing up for convenience.}

\item{rstudio_pref_path}{string: Default rstudio config path.}
}
\value{
Copy rstudio_bindings.json, addins.json rstudio-bindings.json and
r.snippets to .bak in your R Studio config path.
}
\description{
Copy rstudio_bindings.json, addins.json rstudio-bindings.json and r.snippets
to .bak in your R Studio config path.
}
\examples{
backup_prefs(open_backup_path = FALSE)

}
