% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_github.R
\name{import_from_github}
\alias{import_from_github}
\title{Import preferences from a GitHub url}
\usage{
import_from_github(clone_git = FALSE, git_url = NULL, git_path = ".")
}
\arguments{
\item{clone_git}{boolean. Use this option if you haven't initiated a repository.}

\item{git_url}{string. Allow to user the bypass the request of the git url to
directly clone the desired git url repository
@param git_path string. Git path to the repository to pull from.}
}
\value{
Clone or pull from a GitHub link from a user prompt.
}
\description{
Pull (or clone) from your own git repository which
should contain your backup preference files. The function will attempt to pull your current git repository by default.
}
\examples{
\dontrun{
# Pull changes from a GitHub repository
import_from_github()

# Clone a repository from GitHub
import_from_github(clone_git = TRUE)
import_from_github(git_url = "https://github.com/cran/dummies")
}
}
