% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_github.R
\name{import_from_github}
\alias{import_from_github}
\title{Import preferences from a GitHub url}
\usage{
import_from_github(clone_git = FALSE, target_dir = NULL)
}
\arguments{
\item{clone_git}{boolean. Use this option if you haven't initiated a repository.}
}
\value{
Clone or pull from a GitHub link from a user prompt.
}
\description{
Pull (or clone) from your own git repository which
should contain your backup preference files. The function will attempt to pull your current git repository by default.
}
\examples{
\dontshow{if (has_git_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
import_from_github()
import_from_github(clone_git = TRUE)
\dontshow{\}) # examplesIf}
}
