% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_local_prefs.R
\name{import_local_prefs}
\alias{import_local_prefs}
\title{Import local preferences files to R Studio}
\usage{
import_local_prefs(preference_path_name = "rstudio_preferences")
}
\arguments{
\item{preference_path_name}{string: Relative folder_path where .json files are contained.}
}
\value{
Your settings are imported.
}
\description{
Import local preferences files to R Studio
}
\examples{
import_local_prefs()
}
