% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_import_prefs.R
\name{start_backup_prefs}
\alias{start_backup_prefs}
\title{Start the backup process}
\usage{
start_backup_prefs()
}
\value{
Function will start backing up your files in their folder to .bak and
  ask you if you want to back them up to github
}
\description{
Start the backup process
}
\examples{
start_backup_prefs()

}
