% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_import_prefs.R
\name{start_import_prefs}
\alias{start_import_prefs}
\title{Import preferences from GitHub}
\usage{
start_import_prefs(
  pull_github = TRUE,
  clone_git = FALSE,
  preference_path = "/rstudio_preferences"
)
}
\arguments{
\item{pull_github}{boolean. Default is TRUE.}

\item{clone_git}{boolean: Default is FALSE. Clone git for first time usage.}

\item{preference_path}{Input file path where your preferences.json files are
located.}
}
\value{
Pull your github and import your preferences to the R Studio
preference folder.
}
\description{
Import preferences by cloning a link or pulling from your GitHub. The
function will start to import from GitHub, then will import local prefs
}
\examples{
\dontrun{
start_import_prefs(pull_github = FALSE, preference_path = "/rstudio_preferences")
start_import_prefs(pull_github = TRUE)
}

}
