% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_import_prefs.R
\name{start_import_prefs}
\alias{start_import_prefs}
\title{Start the import process from local files or from GitHub}
\usage{
start_import_prefs(
  preference_path = "R/rstudio_preferences/",
  github = TRUE,
  clone_git = FALSE,
  target_dir = NULL
)
}
\arguments{
\item{preference_path}{string: Input the file path where your preferences.json
files are located.}

\item{github}{boolean: Use TRUE to pull or clone from GitHub, otherwise FALSE
to just import your local files.}

\item{clone_git}{boolean: Default is FALSE. Clone git for first time usage. A convenience function to avoid doing it manually, but it's essentially the same as doing \code{git_clone("github_url")}. It will clone the repository inside your current working directory or the specified target directory.}

\item{target_dir}{string: The desired file path to clone your cloned github directory. Use \code{parent_dir <- dirname(getwd())} to get the parent directory of the current directory.}
}
\value{
Pull your github and import your preferences to the R Studio
preference folder.
}
\description{
Import preferences by cloning a link or pulling from your GitHub. The
function will start to pull from the currently active GitHub repository, then will import the local .json
preference files. The function will look for the specific .json files only, in the given file path.
}
\details{
Function wrapper of \code{import_from_github()} followed by \code{import_local_prefs()}
}
\examples{
\dontshow{if (has_git_repository()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# For example when your current rstudio preferences are in the current
working directory and not on github
start_import_prefs(pull_github = FALSE, preference_path = "R/rstudio_preferences/")
# To clone the github containing the rstudio_preferences into the current parent directory of your working directory (you don't want nested github directories)
start_import_prefs(pull_github = TRUE, target_dir = parent_dir <- dirname(getwd()))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=import_from_github]{import_from_github()}} to just import, and \code{\link[=import_local_prefs]{import_local_prefs()}} to import locally
}
