% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_import_prefs.R
\name{start_import_prefs}
\alias{start_import_prefs}
\title{Start import}
\usage{
start_import_prefs(pull_github = TRUE)
}
\arguments{
\item{pull_github}{boolean. Default is TRUE.}
}
\value{
Pull your github and import your preferences to the R Studio preference folder.
}
\description{
Start import
}
\examples{
start_import_prefs()

}
