% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_prefs_github.R
\name{upload_prefs_to_github}
\alias{upload_prefs_to_github}
\title{Upload preferences to GitHub}
\usage{
upload_prefs_to_github(preference_path_name = "rstudio-preferences")
}
\arguments{
\item{preference_path_name}{string. Relative file path name}
}
\value{
Will upload your .json files to your initiated git repository.
}
\description{
Add, commit and push your .json files automatically.
}
\examples{
upload_prefs_to_github()
}
