% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_prefs_github.R
\name{upload_prefs_to_github}
\alias{upload_prefs_to_github}
\title{Upload preferences to GitHub}
\usage{
upload_prefs_to_github(
  preference_path = "R/rstudio_preferences/",
  git_message = "Backup of R Studio preferences on {Sys.Date()}.",
  repository = "."
)
}
\arguments{
\item{preference_path}{string. Relative file path name}

\item{git_message}{string. Enter a git commit message.}

\item{repository}{string. Defaults to the current git repository.}
}
\value{
Upload your .json files to currently active git repository.
}
\description{
Add, commit and push your .json files automatically with the name
}
\examples{

# Setup (inspired from gert)
oldwd <- getwd()
repo <- file.path(tempdir(), "myrepo")
gert::git_init(repo)
setwd(repo)

# Upload preferences to currently active git repository
upload_prefs_to_github("R/rstudio_preferences/")

# Upload preferences to currently active git repository with a custom git commit message
upload_prefs_to_github("R/rstudio_preferences/", git_message = "Backup preferences")

# cleanup
setwd(oldwd)
unlink(repo, recursive = TRUE)

}
